+++ presentation/src/main/java/com/example/presentation/navigation/BottomBarItem.kt
@@ -1,0 +1,7 @@
+package com.example.presentation.navigation
+
+internal data class BottomBarItem(
+    val destination: BottomTabDestinations,
+    val icon: Int,
+    val label: String
+)
